utils/formatting.py
# This file is kept for API compatibility but no longer contains Streamlit-specific code
from PIL import Image
import os

# API formatting functions can be added here if needed in the future
# All Streamlit-specific functionality has been removed

def format_api_response(data):
    """Format API responses consistently"""
    return {
        "success": True,
        "data": data,
        "timestamp": None  # Can be added if timestamp tracking is needed
    }