requirements.txt
numpy
pandas
matplotlib
opencv-python
scikit-learn
torch
google-generativeai
python-dotenv
ultralytics
fastapi
uvicorn
python-multipart
pydantic
pillow